requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
ansible==10.5.0
ansible-core==2.17.6
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Flask==3.0.3
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
meraki==1.51.0
multidict==6.1.0
packaging==24.1
propcache==0.2.0
pycparser==2.22
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
resolvelib==1.0.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.2
yarl==1.17.1
